+++ Source/Guard.cs
@@ -54,11 +54,11 @@ namespace Moq
 		/// Ensures the given <paramref name="value"/> is not null.
 		/// Throws <see cref="ArgumentNullException"/> otherwise.
 		/// </summary>
-		public static void NotNull<T>(Expression<Func<T>> reference, T value)
+		public static void NotNull(object value, string paramName)
 		{
 			if (value == null)
 			{
-				throw new ArgumentNullException(GetParameterName(reference));
+				throw new ArgumentNullException(paramName);
 			}
 		}
 
@@ -67,12 +67,16 @@ namespace Moq
 		/// Throws <see cref="ArgumentNullException"/> in the first case, or 
 		/// <see cref="ArgumentException"/> in the latter.
 		/// </summary>
-		public static void NotNullOrEmpty(Expression<Func<string>> reference, string value)
+		public static void NotNullOrEmpty(string value, string paramName)
 		{
-			NotNull<string>(reference, value);
+			if (value == null)
+			{
+				throw new ArgumentNullException(paramName);
+			}
+
 			if (value.Length == 0)
 			{
-				throw new ArgumentException(Resources.ArgumentCannotBeEmpty, GetParameterName(reference));
+				throw new ArgumentException(Resources.ArgumentCannotBeEmpty, paramName);
 			}
 		}
 
@@ -81,16 +85,16 @@ namespace Moq
 		/// </summary>
 		/// <typeparam name="T">Type of the argument to check, it must be an <see cref="IComparable"/> type.
 		/// </typeparam>
-		/// <param name="reference">The expression containing the name of the argument.</param>
 		/// <param name="value">The argument value to check.</param>
 		/// <param name="from">The minimun allowed value for the argument.</param>
 		/// <param name="to">The maximun allowed value for the argument.</param>
-		public static void NotOutOfRangeInclusive<T>(Expression<Func<T>> reference, T value, T from, T to)
+		/// <param name="paramName">The name of the parameter.</param>
+		public static void NotOutOfRangeInclusive<T>(T value, T from, T to, string paramName)
 				where T : IComparable
 		{
 			if (value != null && (value.CompareTo(from) < 0 || value.CompareTo(to) > 0))
 			{
-				throw new ArgumentOutOfRangeException(GetParameterName(reference));
+				throw new ArgumentOutOfRangeException(paramName);
 			}
 		}
 
@@ -99,20 +103,20 @@ namespace Moq
 		/// </summary>
 		/// <typeparam name="T">Type of the argument to check, it must be an <see cref="IComparable"/> type.
 		/// </typeparam>
-		/// <param name="reference">The expression containing the name of the argument.</param>
 		/// <param name="value">The argument value to check.</param>
 		/// <param name="from">The minimun allowed value for the argument.</param>
 		/// <param name="to">The maximun allowed value for the argument.</param>
-		public static void NotOutOfRangeExclusive<T>(Expression<Func<T>> reference, T value, T from, T to)
+		/// <param name="paramName">The name of the parameter.</param>
+		public static void NotOutOfRangeExclusive<T>(T value, T from, T to, string paramName)
 				where T : IComparable
 		{
 			if (value != null && (value.CompareTo(from) <= 0 || value.CompareTo(to) >= 0))
 			{
-				throw new ArgumentOutOfRangeException(GetParameterName(reference));
+				throw new ArgumentOutOfRangeException(paramName);
 			}
 		}
 
-		public static void CanBeAssigned(Expression<Func<object>> reference, Type typeToAssign, Type targetType)
+		public static void CanBeAssigned(Type typeToAssign, Type targetType, string paramName)
 		{
 			if (!targetType.IsAssignableFrom(typeToAssign))
 			{
@@ -122,21 +126,15 @@ namespace Moq
 						CultureInfo.CurrentCulture,
 						Resources.TypeNotImplementInterface,
 						typeToAssign,
-						targetType), GetParameterName(reference));
+						targetType), paramName);
 				}
 
 				throw new ArgumentException(string.Format(
 					CultureInfo.CurrentCulture,
 					Resources.TypeNotInheritFromType,
 					typeToAssign,
-					targetType), GetParameterName(reference));
+					targetType), paramName);
 			}
-		}
-
-		private static string GetParameterName(LambdaExpression reference)
-		{
-			var member = (MemberExpression)reference.Body;
-			return member.Member.Name;
 		}
 	}
 }